requirements.txt
scrapy>=2.11.0
twisted>=22.10.0
itemadapter>=0.7.0
requests>=2.31.0
urllib3>=2.0.0
python-dotenv>=1.0.0
celery>=5.3.0
redis>=5.0.0
kombu>=5.3.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
flower>=2.0.0
